+++ CHANGELOG.md
@@ -5,6 +5,13 @@ these changes are prefixed with "**BREAKING**"
 
 ## [Unreleased][unreleased]
 
+(no unreleased changes yet)
+
+
+## [v1.2.2][]
+
+**Release date: 2020-03-03**
+
 ### Fixed
 
 * `delphin.ace.ACETransferer` now has a `stderr` parameter ([#278][])
@@ -1185,6 +1192,7 @@ information about changes, except for
 [commit messages](../../commits/v0.2).
 
 [unreleased]: ../../tree/develop
+[v1.2.2]: ../../releases/tag/v1.2.2
 [v1.2.1]: ../../releases/tag/v1.2.1
 [v1.2.0]: ../../releases/tag/v1.2.0
 [v1.1.0]: ../../releases/tag/v1.1.0